// frontend/src/modules/booking/models/reserved_seat.ts
export interface IReservedSeatParam {
    scheduleId: number 
    seatId: number 
    userId: number 
}

export interface IReservedSeat {
    id: number 
    scheduleId: number 
    seatId: number 
    userId: number 
}

export interface IReservedSeatSearch {
    id: number 
    scheduleId: number 
    seatId: number 
    userId: number 
}